Restructured so that checkForCongestionCharge is in its own file but the specific functions for Day Overview and for the tables, that call this, are in their original files
